refactor: streamline repository imports and enhance error handling in transaction use cases
