+++ apps/backend/src/application/usecases/accounts/__tests__/getAllAccounts.test.ts
@@ -1,6 +1,6 @@
 import { CurrencyCode, Money } from '@ledgerly/shared/types';
 import { createUser } from 'src/db/createTestUser';
-import { AccountRepository } from 'src/infrastructure/db/accounts/account.repository';
+import { AccountRepository } from 'src/infrastructure/db/';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { GetAllAccountsUseCase } from '../getAllAccounts';
