+++ apps/backend/src/infrastructure/db/accounts/index.ts
@@ -1,0 +1,1 @@
+export { AccountRepository } from './account.repository';
