+++ apps/backend/src/infrastructure/db/currency/index.ts
@@ -1,0 +1,1 @@
+export { CurrencyRepository } from './currency.repository';
