+++ apps/backend/src/infrastructure/db/entries/index.ts
@@ -1,0 +1,1 @@
+export { EntryRepository } from './entry.repository';
