+++ apps/backend/src/infrastructure/db/index.ts
@@ -1,3 +1,11 @@
 export { TransactionManager } from './TransactionManager';
 export { BaseRepository } from './BaseRepository';
-export { TransactionQueryRepository } from './transaction/transaction-query.repository';
+export {
+  TransactionQueryRepository,
+  TransactionRepository,
+} from './transaction/';
+export { OperationRepository } from './operations';
+export { EntryRepository } from './entries';
+export { AccountRepository } from './accounts';
+export { CurrencyRepository } from './currency';
+export { UserRepository } from './user';
