+++ apps/backend/src/infrastructure/db/operations/index.ts
@@ -1,0 +1,1 @@
+export { OperationRepository } from './operation.repository';
