+++ apps/backend/src/infrastructure/db/transaction/index.ts
@@ -1,0 +1,2 @@
+export { TransactionRepository } from './transaction.repository';
+export { TransactionQueryRepository } from './transaction-query.repository';
