+++ apps/backend/src/infrastructure/db/user/index.ts
@@ -1,0 +1,1 @@
+export { UserRepository } from './user.repository';
